replace old server with new server
